기능 변경 / DC 액비밸브 수용

+++ src/main/java/com/kt/cmmn/util/StringUtil.java
@@ -261,7 +261,7 @@ public final class StringUtil  {
 
     public static String numberFormat( int value )
     {
-        return new java.text.DecimalFormat  (",##0.##").format( (long)value );
+        return new java.text.DecimalFormat  (",##0.##").format(value);
 
     }
